test(generic): relocate tests to `source`
